Error if n < c
Previously if you set n to 50 and concurrency to 100 the program would exit 0
without making any requests. Instead let's print an error message and display
the program's usage.

Happy to add tests but I think this would need a larger refactoring to do flag
parsing outside the main function :( I verified that this works by running it
at the command line with and without n < c.